requirements.txt
python-telegram-bot==22.5
yt-dlp==2025.9.26
python-dotenv==1.1.1